+++ Common_glTF_Exporter/Common_glTF_Exporter.projitems
@@ -20,7 +20,11 @@
     <Compile Include="$(MSBuildThisFileDirectory)ExternalCommand.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core\glTF.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Core\GlTFExportContext.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Model\GeometryDataObject.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Model\PointIntObject.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Model\VertexLookupIntObject.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Utils\SettingsConfig.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Windows\MainWindow\ExportToZero.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Windows\MainWindow\Preferences.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Windows\MainWindow\Settings.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Windows\MainWindow\UpdateForm.cs" />
