+++ Common_glTF_Exporter/Core/Containers.cs
@@ -6,19 +6,6 @@ using Autodesk.Revit.DB;
 namespace Revit_glTF_Exporter
 {
     /// <summary>
-    /// Intermediate data format for 
-    /// converting between Revit Polymesh
-    /// and glTF buffers.
-    /// </summary>
-    public class GeometryData
-    {
-        public List<double> vertices = new List<double>();
-        public List<double> normals = new List<double>();
-        public List<double> uvs = new List<double>();
-        public List<int> faces = new List<int>();
-    }
-
-    /// <summary>
     /// Container for holding a strict set of items
     /// that is also addressable by a unique ID.
     /// </summary>
@@ -119,133 +106,6 @@ namespace Revit_glTF_Exporter
         {
             if (index < 0 || index > List.Count - 1) throw new Exception("Specified item could not be found.");
             return List[index];
-        }
-    }
-
-    /// <summary>
-    /// From Jeremy Tammik's RvtVa3c exporter:
-    /// https://github.com/va3c/RvtVa3c
-    /// An integer-based 3D point class.
-    /// </summary>
-    class PointInt : IComparable<PointInt>
-    {
-        public double X { get; set; }
-        public double Y { get; set; }
-        public double Z { get; set; }
-        private double _boundingBoxMidPointX { get; set; }
-        private double _boundingBoxMidPointY { get; set; }
-        private double _boundingBoxMidPointZ { get; set; }
-
-        public PointInt(XYZ p, bool switch_coordinates,
-
-            #if REVIT2019 || REVIT2020
-
-            DisplayUnitType displayUnitType,
-
-            #else
-
-            ForgeTypeId forgeTypeId,
-
-            #endif
-            bool relocateTo0, XYZ pointToRelocate)
-        {
-            #if REVIT2019 || REVIT2020
-
-            X = Util.ConvertFeetToUnitTypeId(p.X, displayUnitType);
-            Y = Util.ConvertFeetToUnitTypeId(p.Y, displayUnitType);
-            Z = Util.ConvertFeetToUnitTypeId(p.Z, displayUnitType);
-
-            if (relocateTo0)
-            {
-                _boundingBoxMidPointX = Util.ConvertFeetToUnitTypeId(pointToRelocate.X, displayUnitType);
-                _boundingBoxMidPointY = Util.ConvertFeetToUnitTypeId(pointToRelocate.Y, displayUnitType);
-                _boundingBoxMidPointZ = Util.ConvertFeetToUnitTypeId(pointToRelocate.Z, displayUnitType);
-            }
-
-            #else
-
-            X = Util.ConvertFeetToUnitTypeId(p.X, forgeTypeId);
-            Y = Util.ConvertFeetToUnitTypeId(p.Y, forgeTypeId);
-            Z = Util.ConvertFeetToUnitTypeId(p.Z, forgeTypeId);
-
-            if (relocateTo0)
-	        {
-                _boundingBoxMidPointX = Util.ConvertFeetToUnitTypeId(p.X, forgeTypeId);
-                _boundingBoxMidPointY = Util.ConvertFeetToUnitTypeId(p.Y, forgeTypeId);
-                _boundingBoxMidPointZ = Util.ConvertFeetToUnitTypeId(p.Z, forgeTypeId);
-	        }
-
-            #endif
-
-            if (relocateTo0)
-            {
-                pointToRelocate = new XYZ(_boundingBoxMidPointX, _boundingBoxMidPointY, _boundingBoxMidPointZ);
-
-                X -= pointToRelocate.X;
-                Y -= pointToRelocate.Y;
-                Z -= pointToRelocate.Z;
-            }
-
-            if (switch_coordinates)
-            {
-                X = -X;
-                double tmp = Y;
-                Y = Z;
-                Z = tmp;
-            }
-        }
-
-        public int CompareTo(PointInt a)
-        {
-            double d = X - a.X;
-            if (0 == d)
-            {
-                d = Y - a.Y;
-                if (0 == d)
-                {
-                    d = Z - a.Z;
-                }
-            }
-            return (0 == d) ? 0 : ((0 < d) ? 1 : -1);
-        }
-    }
-
-    /// <summary>
-    /// From Jeremy Tammik's RvtVa3c exporter:
-    /// https://github.com/va3c/RvtVa3c
-    /// A vertex lookup class to eliminate 
-    /// duplicate vertex definitions.
-    /// </summary>
-    class VertexLookupInt : Dictionary<PointInt, int>
-    {
-        /// <summary>
-        /// Define equality for integer-based PointInt.
-        /// </summary>
-        class PointIntEqualityComparer : IEqualityComparer<PointInt>
-        {
-            public bool Equals(PointInt p, PointInt q)
-            {
-                return 0 == p.CompareTo(q);
-            }
-
-            public int GetHashCode(PointInt p)
-            {
-                return (p.X.ToString()
-                  + "," + p.Y.ToString()
-                  + "," + p.Z.ToString())
-                  .GetHashCode();
-            }
-        }
-
-        /// <summary>
-        /// Return the index of the given vertex,
-        /// adding a new entry if required.
-        /// </summary>
-        public int AddVertex(PointInt p)
-        {
-            return ContainsKey(p)
-              ? this[p]
-              : this[p] = Count;
         }
     }
 }
